+++ src/sched.c
@@ -68,7 +68,7 @@ void Sched_SetWorkfunc(Task_t tasknum, SchedCall_t func) {
 
 int32_t Sched_Do(uint32_t fastforward) {
 	static uint32_t oldTick = 0;
-	int32_t shortestwait = 0x7fffffff;
+	uint32_t shortestwait = 0x7fffffff;	// max for int32_t!
 	uint32_t curTick = Sched_GetTick();
 
 	// How many ticks will we should roll forward (including sleep time)
@@ -97,7 +97,7 @@ int32_t Sched_Do(uint32_t fastforward) {
 	}
 	// Unless a (wake-up) interrupt calls Sched_SetState, this is how
 	// long it's OK to sleep until next task is due
-	return shortestwait;
+	return (int32_t) shortestwait;
 }
 
 void BusyWait( uint32_t numticks ) {
